Settings: make unaccounted and over-counted battery usage configurable
Google disabled it for release builds here:

https://github.com/SlimRoms/packages_apps_Settings/commit/e6793771cedb47aed72f1c64f870b70357746938

well but it is enabled for whatever other type of builds beside the
official release builds. That said it should
be a developer option so or so and not build type dependent.

Whatever added an option into developer settings to control it. Default
value is false.

That said.....finally no complain about misc battery stat in battery
settings xD

Change-Id: I7549b1c796d5d9413b4af5d80a31d6efeaf0352a

Conflicts:
	res/values/slim_strings.xml
	src/com/android/settings/DevelopmentSettings.java

Use checkbox not Switchpreference

(want to switch to Switches, need to change or add for SwitchPreference:

private CheckBoxPreference findAndInitCheckboxPref(String key) {
        CheckBoxPreference pref = (CheckBoxPreference)
findPreference(key);
        if (pref == null) {
            throw new IllegalArgumentException("Cannot find preference
with key = " + key);
        }
        mAllPrefs.add(pref);
        mResetCbPrefs.add(pref);
        return pref;
    }

)

- This is the 2nd commit message:

set unaccounted and over-counted battery setting as Checkbox